backend/pyproject.toml
[build-system]
requires = ["setuptools>=75"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = []

[project]
name = "init-backend"
version = "0.0.1"
requires-python = ">=3.12"
readme = "README.md"

dependencies = [
    # Web framework
    "fastapi==0.116.1",
    "uvicorn==0.35.0",
    "starlette==0.47.2",

    # HTTP requests & networking
    "httpx==0.28.1",
    "requests==2.32.4",

    # Utilities
    "python-dotenv==1.1.1",

    # Database
    "sqlalchemy[asyncio]==2.0.30",
    "asyncpg==0.29.0",
    "alembic==1.13.1",
]

[project.optional-dependencies]
dev = [
    "pre-commit",
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "pytest-env",
    "pytest-freezegun",
    "setuptools",
]

deploy = [
    "build",
    "twine",
]
